fix `checkOwner`

+++ src/app/wallet/services/wallet.service.ts
@@ -45,10 +45,13 @@ export class WalletService {
   }
 
   async checkOwner(): Promise<any> {
-    const owner = await this.walletSC['owner']();
-    if (owner.toLowerCase() !== this.walletOwner.address.toLowerCase()) {
-      localStorage.clear();
-      this.snackBar.open("You are not the owner anymore.")
+    const isDeployed = await this.isDeployed();
+    if (isDeployed === WalletState.READY) {
+      const owner = await this.walletSC['owner']();
+      if (owner.toLowerCase() !== this.walletOwner.address.toLowerCase()) {
+        localStorage.clear();
+        this.snackBar.open("You are not the owner anymore.")
+      }
     }
   }
